fix: exports block and image hooks and getters
- exports getPrompt getter
- correct response types for image, prompt and block
- consistent response naming
- abstracts image types to types folder
- adds 'use client' directive to all hooks
