+++ libs/future-components/src/client/index.ts
@@ -3,6 +3,8 @@ export { usePrompt } from '../handlers/prompt/usePrompt'
 export { useRequest } from '../hooks/useRequest'
 export { useErrorEnhancement } from '../handlers/errorEnhancer/useErrorEnhancement'
 export { useSelect } from '../handlers/select/useSelect'
+export { useBlock } from '../handlers/block/useBlock'
+export { useImage } from '../handlers/image/useImage'
 export { useText } from '../handlers/text/useText'
 export { renderToString } from './renderToString'
 export { Block } from '../components/Block'
