+++ libs/future-components/src/components/Image.tsx
@@ -1,94 +1,11 @@
-import NextImage, { type ImageProps as NextImageProps } from 'next/image'
+import NextImage from 'next/image'
+import { getImage } from '../handlers/image/getters'
 import type { SyntheticEvent } from 'react'
-import { request } from '../utils/request'
-import { IS_DEV } from '../utils/constants'
-import { ApiUrlEnum } from '../enums/ApiUrlEnum'
-import type { ImageRequestBody } from '../handlers/image/models'
-
-export interface ImageDescribeProps
-	extends Omit<NextImageProps, 'src' | 'alt' | 'onLoad' | 'onError'> {
-	/**
-	 * URL to the image to describe. Generates `alt` text.
-	 * The `alt` text is then passed to the image and returned in `onLoad`
-	 */
-	src: string
-	/**
-	 * Whether or not to render the result string adjacent to the image
-	 */
-	showResult?: boolean
-	/**
-	 * Callback function invoked once the image is completely loaded and the `placeholder` has been removed.
-	 * Returns the `event`, description `response` and the original `src`
-	 */
-	onLoad?: ImageDescribeCallback
-	/** Callback function that is invoked if the image fails to load. */
-	onError?: ImageDescribeCallback
-}
-
-export type ImageDescribeCallback = (
-	event: SyntheticEvent<HTMLImageElement, Event> | undefined,
-	/** Image description response */
-	response?: string,
-	/** Image description src */
-	src?: string
-) => void
-
-export interface ImageGenerateProps
-	extends Omit<
-		NextImageProps,
-		'src' | 'width' | 'height' | 'alt' | 'onLoad' | 'onError'
-	> {
-	/** Image generation prompt */
-	prompt: string
-	/**
-	 * Whether or not to render the result string adjacent to the image
-	 */
-	showResult?: boolean
-	/** Width and height of the image. Default '256x256' */
-	size?: '256x256' | '512x512' | '1024x1024'
-	/** Alternative text describing the image, uses `prompt` if not provided */
-	alt?: string
-	/**
-	 * Callback function invoked once the image is completely loaded and the `placeholder` has been removed.
-	 * Returns the `event`, generation result `response` and the original `prompt`
-	 */
-	onLoad?: ImageGenerateCallback
-	/** Callback function that is invoked if the image fails to load. */
-	onError?: ImageGenerateCallback
-}
-
-export type ImageGenerateCallback = (
-	event: SyntheticEvent<HTMLImageElement, Event> | undefined,
-	/** Image generation response */
-	response?: string,
-	/** Image generation prompt */
-	prompt?: string
-) => void
-
-export type ImageProps<T> = T extends { prompt: unknown }
-	? ImageGenerateProps
-	: T extends { alt: unknown }
-	? NextImageProps
-	: ImageDescribeProps
-
-/**
- * Image generation and description fetcher
- */
-export function getImage<T>(props: ImageProps<T>) {
-	const { prompt, src } = props as ImageRequestBody
-
-	if (!('prompt' in props) && 'alt' in props) {
-		if (IS_DEV) {
-			console.warn(
-				`Provided 'alt' and no 'prompt' for image.\nYou may want to use '<Image>' from 'next/image' instead.`
-			)
-		}
-
-		return false
-	}
-
-	return request<string>(ApiUrlEnum.image, { body: { prompt, src } })
-}
+import type {
+	ImageProps,
+	ImageDescribeCallback,
+	ImageGenerateCallback,
+} from '../types/Image'
 
 /**
  * client-only callback for Image `onLoad` | `onError`
