+++ libs/future-components/src/handlers/notFoundEnhancer/useNotFoundEnhancement.ts
@@ -1,3 +1,5 @@
+'use client'
+
 import type {
 	NotFoundEnhancerResponse,
 	NotFoundEnhancerRequestBody,
