+++ libs/future-components/src/handlers/prompt/usePrompt.ts
@@ -1,10 +1,13 @@
+'use client'
+
+import type { PromptResponse } from './models'
 import {
 	useRequest,
 	type UseRequestConsumerConfig,
 } from '../../hooks/useRequest'
 import { ApiUrlEnum } from '../../enums/ApiUrlEnum'
 
-export const usePrompt = <T = string>(
+export const usePrompt = <T = PromptResponse>(
 	prompt: string,
 	config?: UseRequestConsumerConfig<T>
 ) => {
