+++ libs/future-components/src/handlers/select/useSelect.ts
@@ -1,3 +1,5 @@
+'use client'
+
 import type { SelectResponse, SelectRequestBody } from './models'
 import {
 	useRequest,
