+++ libs/future-components/src/handlers/text/useText.tsx
@@ -1,4 +1,5 @@
 'use client'
+
 import {
 	useRequest,
 	type UseRequestConsumerConfig,
