+++ libs/future-components/src/hooks/useRequest.ts
@@ -1,4 +1,5 @@
 'use client'
+
 import { useState, useEffect, useRef } from 'react'
 import { request } from '../utils/request'
 import type { RequestConfig } from '../types'
