+++ libs/future-components/src/index.ts
@@ -69,6 +69,9 @@ export { Image } from './components/Image'
 export { Select } from './components/Select'
 
 // Service utils
+export { getPrompt } from './handlers/prompt/getters'
+export { getBlock } from './handlers/block/getters'
+export { getImage } from './handlers/image/getters'
 export { getSelect } from './handlers/select/getters'
 export { getText } from './handlers/text/getters'
